==> 경찰대_체감안전도예측/00.data_install.ipynb <==
mkdir data

from geoband.API import *
GetCompasData('SBJ_2107_004', '1', '1.5개시도_경찰서_관할경계.geojson')
GetCompasData('SBJ_2107_004', '2', '2.5개시도_치안안전센터_관할경계.geojson')
GetCompasData('SBJ_2107_004', '3', '3.체감안전도_설문지.txt')
GetCompasData('SBJ_2107_004', '4', '4.체감안전도_통계(2017~2019).csv')
GetCompasData('SBJ_2107_004', '5', '5.112신고.csv')
GetCompasData('SBJ_2107_004', '6', '6.범죄발생원표.csv')
GetCompasData('SBJ_2107_004', '7', '7.범죄검거원표.csv')
GetCompasData('SBJ_2107_004', '8', '8.범죄피의자원표.csv')
GetCompasData('SBJ_2107_004', '9', '9.5개시도_보안등현황.csv')
GetCompasData('SBJ_2107_004', '10', '10.5개시도_CCTV설치현황.csv')
GetCompasData('SBJ_2107_004', '11', '11.5개시도_교통사고현황.csv')
GetCompasData('SBJ_2107_004', '12', '12.경찰서별_지구대별 인원현황.csv')
GetCompasData('SBJ_2107_004', '13', '13.5개시도_화재발생통계.csv')
GetCompasData('SBJ_2107_004', '14', '14.5개시도_성연령별_총인구분포.csv')
GetCompasData('SBJ_2107_004', '15', '15.5개시도_1인가구수.csv')
GetCompasData('SBJ_2107_004', '16', '16.5개시도_인구데이터(외국인).csv')
GetCompasData('SBJ_2107_004', '17', '17.5개시도_기초수급자현황.csv')
GetCompasData('SBJ_2107_004', '18', '18.5개시도_자살율.csv')
GetCompasData('SBJ_2107_004', '19', '19.5개시도_최종학력통계.csv')
GetCompasData('SBJ_2107_004', '20', '20.5개시도_공원현황.csv')
GetCompasData('SBJ_2107_004', '21', '21.5개시도_유흥업소_단란주점현황.csv')
GetCompasData('SBJ_2107_004', '22', '22.5개시도_비상벨현황.geojson')
GetCompasData('SBJ_2107_004', '23', '23.5개시도_종관관측기상(2017~2020).csv')
GetCompasData('SBJ_2107_004', '24', '24.5개시도_도로명주소(건물).geojson')
GetCompasData('SBJ_2107_004', '25', '25.결과제출양식.csv')
GetCompasData('SBJ_2107_004', '26', '26.코드정의서.xlsx')

!mv *'.csv' ./data
!mv *'.geojson' ./data
!mv *'.xlsx' ./data
!mv *'.txt' ./data
