==> 경찰대_체감안전도예측/1.2.Q2_강도살인안전도_final.ipynb <==
import pandas as pd

## Data Load

q1_total = pd.read_csv('q1_total.csv')
del q1_total['Unnamed: 0']
q1_total.head()

report_112 = pd.read_csv('report_112.csv')
del report_112['Unnamed: 0']
report_112.head()

q1_total.columns

# 강도살인에서 쓸 데이터만 남김
q2_info = q1_total[['jur_stn', 'date', 'pub_cnt', 'report_total_cnt', 'report_sx_남성',
       'report_sx_불상', 'report_sx_여성', 'poor_num', 'poor_ratio', 'age_합계', 'age_15세미만',
       'age_15~64세', 'age_65세이상', 'age_평균연령', 'sx_rate_합계', 'sx_rate_15세미만',
       'sx_rate_15~64세', 'sx_rate_65세이상', 'foreign_num', 'cctv_cnt', 'bell_cnt',
       'light_cnt', 'mur_safety_q2']]
q2_info.head()

# ---------------------------------------

## ※ 목표 : Q2 X값 생성

- 전체 112 신고건과 강도, 살인 관련 신고건을 count하여 x값으로 사용

report_112.head(3) # Q2에 사용할 data 

### 1. 강도

# 강도 total count
robb_total_cnt = report_112.groupby(['jur_stn','date','inc_info']).count().unstack(fill_value=0).stack()[['report_sx']]
robb_total_cnt.reset_index(inplace=True)
robb_total_cnt = robb_total_cnt.rename(columns={'report_sx' : 'robb_total_cnt'})
robb_total_cnt = robb_total_cnt[robb_total_cnt['inc_info']=='강도']
robb_total_cnt.head()

robb_total_cnt['robb_total_cnt'] = robb_total_cnt.apply(lambda x : x['robb_total_cnt'] * 5 if x['date'] =='2018(상반기)' else x['robb_total_cnt'] , axis= 1   )
# 0 일경우 1로 대체
robb_total_cnt['robb_total_cnt'] = robb_total_cnt.apply(lambda x : 1 if x['robb_total_cnt'] ==0 else x['robb_total_cnt'] , axis= 1   )

del robb_total_cnt['inc_info']
robb_total_cnt.head(3)

### 2. 살인

# 살인 신고자 total count
mur_total_cnt = report_112.groupby(['jur_stn','date','inc_info']).count().unstack(fill_value=0).stack()[['year']]
mur_total_cnt.reset_index(inplace=True)
mur_total_cnt = mur_total_cnt.rename(columns={'year' : 'mur_total_cnt'})
mur_total_cnt = mur_total_cnt[mur_total_cnt['inc_info']=='살인']
mur_total_cnt.head()

mur_total_cnt['mur_total_cnt'] = mur_total_cnt.apply(lambda x : x['mur_total_cnt'] * 5 if x['date'] =='2018(상반기)' else x['mur_total_cnt'] , axis= 1   )
# 0 일경우 1로 대체
mur_total_cnt['mur_total_cnt'] = mur_total_cnt.apply(lambda x : 1 if x['mur_total_cnt'] ==0 else x['mur_total_cnt'] , axis= 1   )

### 1, 2 merge 

# 강도 + 살인 merge
temp_df = pd.merge(robb_total_cnt, mur_total_cnt)
temp_df.head()

## 3. X값 최종 Merge

# feature로 사용될 데이터 종합
q2_df = pd.merge( temp_df, q2_info, how='outer')
q2_df.head()

del q2_df['inc_info']

q2_df['date'].value_counts()

q2_df[q2_df['jur_stn'].str.contains('세종')]

## 세종 null 값 index 채우기 ##
null_index= q2_df[q2_df['cctv_cnt'].isnull()].index 
sejong_mean = q2_df[(q2_df['jur_stn'].str.contains('세종')) & (q2_df['robb_total_cnt'].notnull()) ].mean()

for i in null_index :
    for c in q2_df.columns[2:-1] :
#         print(i, c)
#         print(q1_total.loc[i, c])
        q2_df.loc[i, c] = sejong_mean[c]

q2_df['robb_total_cnt'].fillna(1, inplace=True)
q2_df['mur_total_cnt'].fillna(1, inplace=True)

q2_df[q2_df['jur_stn'].str.contains('세종')]

q2_df.isnull().sum()

q2_df.to_csv('q2_df.csv')

## X_data, y_data 

# q2 data에서 X,y data 분리
X_data = q2_df.drop('mur_safety_q2',axis=1)
y_data = q2_df['mur_safety_q2'].dropna()

# 2020 년도는 y값이 없으므로 제외하고 eval data로 포함
X_data = X_data.drop(X_data[X_data['date'].str.contains('2020')].index)
X_data.head()

# 2020년 데이터로 eval data 생성
del_idx = q2_df[q2_df['mur_safety_q2'].notnull()].index
eval_X_data = q2_df.drop(del_idx)
eval_X_data=  eval_X_data.drop('mur_safety_q2',axis=1)

X_data['date'].value_counts()

eval_X_data['date'].value_counts()

X_data.shape, y_data.shape

eval_X_data.shape

# train data save
eval_X_data.reset_index(drop=True,inplace=True)
train_data = pd.concat([X_data,y_data],axis=1)
train_data.to_csv("./analysis/q2_train.csv",index=0)
eval_X_data.to_csv("./analysis/q2_target.csv",index=0)
