==> 경찰대_체감안전도예측/1.4.Q4_법질서준수도_final.ipynb <==
import pandas as pd

## Data Load

q1_total = pd.read_csv('q1_total.csv')
del q1_total['Unnamed: 0']
q1_total.head()

q1_total.columns

# 법질서준수도에서 쓸 데이터만 남김
q4_info = q1_total[['jur_stn', 'date', 'pub_cnt', 'poor_num', 'age_합계', 'age_15세미만',
       'age_15~64세', 'age_65세이상', 'age_평균연령', 'sx_rate_합계', 'sx_rate_15세미만',
       'sx_rate_15~64세', 'sx_rate_65세이상', 'foreign_num', 'cctv_cnt', 'bell_cnt',
       'light_cnt', 'raw_odder_q4']]
q4_info.head()

### 전국112신고접수 데이터 출처 : 스마트치안빅데이터플랫폼 '전국112신고접수' (경찰청 제공)
- 링크 : https://www.bigdata-policing.kr/policy/contents/policy-dataSearch.do?schM=View&dbiIdx=23

report_data = pd.read_csv('./data/전국112신고접수.csv')
del report_data['Unnamed: 0']
report_data.head()

# -------------------------------------------------

## Data Check

report_data.info()

# 발생시간대 재분류
report_data['새벽(01~05)'] = report_data['발생시간대(01)'] + report_data['발생시간대(05)']
report_data['아침(06~10)'] = report_data['발생시간대(06)'] + report_data['발생시간대(10)']
report_data['점심(11~14)'] = report_data['발생시간대(11)'] + report_data['발생시간대(14)']
report_data['오후(15~18)'] =report_data['발생시간대(15)'] + report_data['발생시간대(18)']
report_data['저녁(19~21)'] = report_data['발생시간대(19)'] + report_data['발생시간대(21)']
report_data['심야(22~00)'] = report_data['발생시간대(22)'] + report_data['발생시간대(00)']

report_data.columns

report_data.drop(['발생시간대(00)', '발생시간대(01)', '발생시간대(02)',
       '발생시간대(03)', '발생시간대(04)', '발생시간대(05)', '발생시간대(06)', '발생시간대(07)',
       '발생시간대(08)', '발생시간대(09)', '발생시간대(10)', '발생시간대(11)', '발생시간대(12)',
       '발생시간대(13)', '발생시간대(14)', '발생시간대(15)', '발생시간대(16)', '발생시간대(17)',
       '발생시간대(18)', '발생시간대(19)', '발생시간대(20)', '발생시간대(21)', '발생시간대(22)',
       '발생시간대(23)'], axis=1, inplace=True) # 기존 컬럼 drop

report_data['date'] = report_data['접수년월'].apply(lambda x : str(x)[:4] + '(상반기)' if  int(str(x)[-2:]) < 7 else str(x)[:4] + '(하반기)' )
report_data = report_data.rename(columns={'경찰서명' : 'jur_stn'})
report_data.head() # date, jur_stn 양식 맞춰줌

### Q4 X값 생성
- 기존 report data를 상, 하반기 sum 값으로 묶어 X값으로 사용

report_data_df = report_data.groupby(['jur_stn','date']).sum()
report_data_df.reset_index(inplace=True)
report_data_df.head(3) # 상, 하반기별 총합 계산

report_data_df.drop('접수년월', axis=1, inplace=True) # 접수년월 col 제외

report_data_df['date'].value_counts()

## 최종 Merge

# feature로 사용될 데이터 종합
law_df = pd.merge(report_data_df, q4_info, on=['jur_stn','date'], how='right')
law_df.head(3)

law_df['date'].value_counts()

law_df[law_df['신고 건수(전체)'].isnull()] # null 값 확인

## 세종 null 값 index 채우기 ##
null_index= law_df[law_df['신고 건수(전체)'].isnull()].index 
sejong_mean = law_df[(law_df['jur_stn'].str.contains('세종')) & (law_df['신고 건수(전체)'].notnull()) ].mean()
law_df.columns

law_df.columns[2:-1]

for i in null_index :
    for c in law_df.columns[2:-5] :
#         print(i, c)
#         print(q1_total.loc[i, c])
        law_df.loc[i, c] = sejong_mean[c]

law_df[law_df['jur_stn'].str.contains('세종')]

law_df.columns

law_df['report_ratio'] = round(law_df['신고 건수(전체)'] / law_df['age_합계'], 2)
law_df.head() # 인구대비 신고율

q4_df = law_df.copy()
q4_df['date'].value_counts()

q4_df.to_csv('q4_df.csv') # 최종

## X , y data

# q4 data에서 X,y data 분리
X_data = q4_df.drop('raw_odder_q4',axis=1)
y_data = q4_df['raw_odder_q4'].dropna()

# 2020 년도는 y값이 없으므로 제외하고 eval data로 포함
X_data = X_data.drop(X_data[X_data['date'].str.contains('2020')].index)
X_data.head()

# 2020년 데이터로 eval data 생성
del_idx = q4_df[q4_df['raw_odder_q4'].notnull()].index
eval_X_data = q4_df.drop(del_idx)
eval_X_data=  eval_X_data.drop('raw_odder_q4',axis=1)

X_data['date'].value_counts()

eval_X_data['date'].value_counts()

X_data.shape, y_data.shape

eval_X_data.shape

# train data save
eval_X_data.reset_index(drop=True,inplace=True)
train_data = pd.concat([X_data,y_data.dropna()],axis=1)
train_data.to_csv("./analysis/q4_train.csv",index=0)
eval_X_data.to_csv("./analysis/q4_target.csv",index=0)
