==> 경찰대_체감안전도예측/1.5.Q5_전반적안전도_final.ipynb <==
import pandas as pd

## Data Load

## 1. 절도폭력 + 인구종합

q1_total = pd.read_csv('q1_total.csv')
del q1_total['Unnamed: 0']
q1_total.head()

q1_total.columns

# 전반적안전도에서 쓸 데이터 남김
q5_info = q1_total[['jur_stn', 'date', 'pub_cnt', 'vio_total_cnt', 'report_sx_남성_x',
       'report_sx_불상_x', 'report_sx_여성_x', 'rob_total_cnt', 'report_sx_남성_y',
       'report_sx_불상_y', 'report_sx_여성_y', 'report_total_cnt', 'report_sx_남성',
       'report_sx_불상', 'report_sx_여성', 'poor_num', 'age_합계', 'age_15세미만',
       'age_15~64세', 'age_65세이상', 'age_평균연령', 'sx_rate_합계', 'sx_rate_15세미만',
       'sx_rate_15~64세', 'sx_rate_65세이상', 'foreign_num', 'cctv_cnt', 'bell_cnt',
       'light_cnt', 'overall_q5']]
q5_info.head()

q2_df = pd.read_csv('q2_df.csv')
del q2_df['Unnamed: 0']
q2_df.head()

# -----------------------------------------------------------

## 2. 강도살인

q2_df.columns

q2_df.drop(['pub_cnt', 'report_total_cnt', 'report_sx_남성',
       'report_sx_불상', 'report_sx_여성', 'poor_num', 'foreign_num', 'cctv_cnt', 'bell_cnt',
       'light_cnt', 'mur_safety_q2', 'robb_total_cnt', 'mur_total_cnt'],axis=1,inplace=True) # 겹치는 column 제외

crime_df = pd.merge(q5_info, q2_df) # q1, q2 
crime_df.head()

crime_df['date'].value_counts()

# -----------------------------------------------------------

## 3. 교통

q3_df  = pd.read_csv('q3_df.csv')
del q3_df['Unnamed: 0']
q3_df.head()

q3_df.columns

q3_df = q3_df[['jur_stn', 'date', 'total_cnt', 'ofn_sx_기타불명', 'ofn_sx_남', 'ofn_sx_여',
       'ofn_sx_성비', 'ofn_age_10대 이하', 'ofn_age_20~30대', 'ofn_age_40~50대',
       'ofn_age_60~70대', 'ofn_age_80대 이상', 'ofn_age_미분류', 'vic_sx_기타불명',
       'vic_sx_남', 'vic_sx_여', 'vic_sx_성비', 'vic_age_10대 이하', 'vic_age_20~30대',
       'vic_age_40~50대', 'vic_age_60~70대', 'vic_age_80대 이상', 'vic_age_미분류',
       'month_1m', 'month_2m', 'month_3m', 'month_4m', 'month_5m', 'month_6m',
       'acc_type_차대사람', 'acc_type_차대차', 'acc_type_차량단독', 'acc_case_공작물충돌',
       'acc_case_기타', 'acc_case_길가장자리구역통행중', 'acc_case_도로외이탈',
       'acc_case_보도통행중', 'acc_case_전도전복', 'acc_case_정면충돌',
       'acc_case_주/정차차량 충돌', 'acc_case_차도통행중', 'acc_case_추돌', 'acc_case_측면충돌',
       'acc_case_횡단중', 'acc_case_후진중충돌']] # 겹치는 column 제외

# -----------------------------------------------------------

## 4. 법질서 준수

q4_df = pd.read_csv('q4_df.csv')
del q4_df['Unnamed: 0']
q4_df.head()

q4_df.columns

q4_df = q4_df[['jur_stn', 'date','접수긴급유형(긴급)', '접수긴급유형(중요)', '접수긴급유형(일반)', '접수긴급유형(민원)',
       '접수긴급유형(기타)', '발생요일(일)', '발생요일(월)', '발생요일(화)', '발생요일(수)', '발생요일(목)',
       '발생요일(금)', '발생요일(토)', '새벽(01~05)', '아침(06~10)', '점심(11~14)',
       '오후(15~18)', '저녁(19~21)', '심야(22~00)']] # 겹치는 column 제외

# -----------------------------------------------------------

## 목표 : Q5 X값 생성
- Q5 전반적안전도는 Q1~Q4를 feature 값으로 사용한다.

### 1 ~ 4 Merge

## Q 3, 4. merge
acc_law_df = pd.merge(q3_df, q4_df, on=['jur_stn','date'])
acc_law_df.head()

# 1/2 , 3/4 merge
q5_df = pd.merge(crime_df, acc_law_df, on=['jur_stn','date'])
q5_df.head()

q5_df['date'].value_counts()

q5_df[q5_df['date'].str.contains('2020')].isnull().sum().sum()

q5_df.isnull().sum().sum() # 82개(2020년도만) 맞는지 check

q5_df.columns

q5_df.head()[['poor_num', 'age_합계', 'age_15세미만',
       'age_15~64세', 'age_65세이상', 'age_평균연령', 'sx_rate_합계', 'sx_rate_15세미만',
       'sx_rate_15~64세', 'sx_rate_65세이상', 'foreign_num', 'cctv_cnt',
       'bell_cnt', 'light_cnt', 'overall_q5']]

## X, y data

# q5 data에서 X,y data 분리
X_data = q5_df.drop('overall_q5',axis=1)
y_data = q5_df['overall_q5'].dropna()

# 2020 년도는 y값이 없으므로 제외하고 eval data로 포함
X_data = X_data.drop(X_data[X_data['date'].str.contains('2020')].index)
X_data.head()

# 2020년 데이터로 eval data 생성
del_idx = q5_df[q5_df['overall_q5'].notnull()].index
eval_X_data = q5_df.drop(del_idx)
eval_X_data=  eval_X_data.drop('overall_q5',axis=1)

X_data['date'].value_counts()

eval_X_data['date'].value_counts()

X_data.shape, y_data.shape

eval_X_data.shape

# train data save
eval_X_data.reset_index(drop=True,inplace=True)
train_data = pd.concat([X_data,y_data.dropna()],axis=1)
train_data.to_csv("./analysis/q5_train.csv",index=0)
eval_X_data.to_csv("./analysis/q5_target.csv",index=0)
