==> 경찰대_체감안전도예측/2.6.make_it_submission.ipynb <==
import pandas as pd
import numpy as np
import warnings as w
w.filterwarnings('ignore')

collist = [1.012,0.944156,1.021942,0.964156,1.068787,1.047006]
collist2 = [11,17,16,22,32,34]
answer = pd.read_csv("../data/4.체감안전도_통계(2017~2019).csv")
sub = pd.read_csv("../data/25.결과제출양식.csv")
pol_order = list(sub['경찰서명'])

for i in range(len(pol_order)):
    pol_order[i] = "경찰서명_"+pol_order[i]
    pol_order[i] = pol_order[i].replace(" ","")

haban19 = answer[answer['large_category']=='2019 하반기(경찰서별)']
sangban19 = answer[answer['large_category']=='2019 상반기(경찰서별)']
haban18 = answer[answer['large_category']=='2018 하반기(경찰서별)']
haban19.reset_index(drop=True,inplace=True)
sangban19.reset_index(drop=True,inplace=True)
haban18.reset_index(drop=True,inplace=True)

result1 = pd.read_csv('q1_result.csv',index_col=0)#['pred']
result2 = pd.read_csv('q2_result.csv',index_col=0)#['0']
result3 = pd.read_csv('q3_result.csv',index_col=0)#['0']
result4 = pd.read_csv('q4_result.csv',index_col=0)#['0']
result5 = pd.read_csv('q5_result.csv',index_col=0)#['0']

sub['문항1. 절도 폭력안전도\n (문항1 점수 * 10)'] = result1
sub['문항2. 강도 살인 안전도\n(문항2 점수 * 10)'] = result2
sub['문항3. 교통사고 안전도\n(문항3 점수*10)'] = result3
sub['문항4. 법질서 준수도\n(문항 4점수 * 10)'] = result4
sub['문항5. 전반적 안전도\n(문항 5점수 * 10)'] = result5

sub[sub.columns[4]] = (sub[sub.columns[5]] * 0.5) + (sub[sub.columns[6]] * 0.5)
sub[sub.columns[3]] = (sub[sub.columns[4]] * 0.343) + (sub[sub.columns[-3]] * 0.305) + (sub[sub.columns[-2]] * 0.352)
sub[sub.columns[2]] = (sub[sub.columns[3]] * 0.7) + (sub[sub.columns[9]] * 0.3)

sub.index = sub['순번']
del sub['순번']

sub[sub.columns[1]] = (sub[sub.columns[2]] * 0.7+0.058125) + (sub[sub.columns[8]] * (0.3+0.005))
for j in range(len(collist)):
    print(collist[j])
    sub.loc[collist2[j],sub.columns[1]]*= collist[j]

sub.to_csv("../submission/final_result.csv",encoding='utf-8-sig')
